grammar, ambiguity, etc
a few grammar tweaks that i think will make thinga a bit easier to read. i also put the first mission's ip , usr, and pass in the first email, and changed the ip to something that conforms to internet protocol version 4. added capital letters and underscore to approved letters for username as well.