Way too much stuff.
grammar, spelling, ambiguity, putting the ip in the mail instead of the text file, and...just a lot of spelling and grammar. hot sure if everything  is 100%, and it's possible that me accidentally running the file could have corrupted it, but i just need to get this commit out and go to bed.